requirements/base.txt
mysqlclient==1.3.10
Django==1.11.5
djangorestframework==3.6.4
django-extensions==1.9.1
django-webpack-loader==0.5.0
ipdb==0.10.3
